+++ GoReceitas/Extensions/NotificationName+.swift
@@ -1,0 +1,12 @@
+//
+//  NotificationName+.swift
+//  GoReceitas
+//
+//  Created by Igor Fernandes on 30/12/22.
+//
+
+import UIKit
+
+extension Notification.Name {
+    static let favoritesUpdated = Notification.Name("favoritesUpdated")
+}
